+++ js/main.js
@@ -6,7 +6,7 @@ import './upload-file.js';
 import {getData} from './api.js';
 import {fillPictureContainer} from './miniature.js';
 import {getPictureClick} from './popup.js';
-import {showAlert} from './data.js';
+import {showAlert} from './util.js';
 import {showSuccess} from './success.js';
 import {setUserFormSubmit} from './user-form.js';
 getData()
